refactor(convertors): split into json/table/tree/metrics modules; add unit tests
